test(control-plane): separate orchestration tests from EC2 provider tests
